third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/max-length-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  const DEVICE_NAME = 'a'.;

  return setUpPreconnectedDevice({name: DEVICE_NAME})
    .then(() => requestDeviceWithKeyDown({ filters: [{name: DEVICE_NAME}]}))
    .then(device => assert_equals(device.name, DEVICE_NAME));
}, 'A device name of 248 bytes is valid.');
</script>
